fix: Widen suppression counter so deep @_ recursion can't overflow it
`VM::suppress_depth` and its checkpoint snapshot were `u16`. Deep `@_`
(suppressive-capture) recursion increments the counter once per open scope, so
a ~65 KB source overflowed it at 65_536 under default limits — a debug panic, a
release wrap that could unsuppress effects or underflow `SuppressEnd`. It is
reachable from a valid query and source, so it broke the trust boundary.

Widen to `u64`: each open scope lives in an active call frame, so the counter
is bounded by call-nesting depth, itself capped by the u32-indexed frame arena —
`u64` cannot overflow first.

Pin it with a type-width regression (the counter must outrange `u16`) and a
recursive `@_` template in `fuzz_no_panic` (the fixed templates had none). A full
>65_536 execution is quadratic in depth, too slow for CI; the width argument plus
the manual repro cover it. Amend AGENTS.md so the safety enumeration also covers
input-scaling integer counters, not just native recursion and heap.

+++ crates/plotnik-vm/src/engine/checkpoint.rs
@@ -37,8 +37,8 @@ pub(crate) struct CheckpointState {
     pub(crate) frame_index: Option<u32>,
     /// Recursion depth at checkpoint.
     pub(crate) recursion_depth: u32,
-    /// Suppression depth at checkpoint.
-    pub(crate) suppress_depth: u16,
+    /// Suppression depth at checkpoint (see `VM::suppress_depth` for its bound).
+    pub(crate) suppress_depth: u64,
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -99,7 +99,7 @@ impl Checkpoint {
     pub fn ip(&self) -> u16 {
         self.ip
     }
-    pub fn suppress_depth(&self) -> u16 {
+    pub fn suppress_depth(&self) -> u64 {
         self.state.suppress_depth
     }
 }
